Update discord.s/builders to 0.12.0, move management commands to another command that is not enabled for @everyone by default, merge set_channel and remove_channel
