+++ package.json
@@ -11,7 +11,7 @@
   "author": "DEF",
   "license": "ISC",
   "dependencies": {
-    "@discordjs/builders": "^0.6.0",
+    "@discordjs/builders": "0.12.0",
     "beautify.log": "^3.0.4",
     "bufferutil": "^4.0.6",
     "chalk": "4",
