+++ src/botconfig/config.json
@@ -1,12 +1,21 @@
 {
   "token": "",
   "bot_name": "Suggestions",
-  "loadSlashsGlobal": true, "_COMMENT_": "Set it to true, when your Slash Commands are ready, false loads instant true loads slow but stable! (GLOBAL/GUILD COMMANDS, DISCORD DOCS!)",
+  "loadSlashsGlobal": false, "_COMMENT_": "Set it to true, when your Slash Commands are ready, false loads instant true loads slow but stable! (GLOBAL/GUILD COMMANDS, DISCORD DOCS!)",
   "slashCommandsDirs": [
         {
                 "Folder": "suggestions",
                 "CmdName": "suggestions",
-                "CmdDescription": "Suggestions"
+                "CmdDescription": "Suggestions",
+                "dmpermission": false,
+                "activeByDefault": true
+        },
+        {
+                "Folder": "manage",
+                "CmdName": "manage",
+                "CmdDescription": "Manage Suggestions",
+                "dmpermission": false,
+                "activeByDefault": false
         }
    ]
 }
