+++ src/slashCommands/manage/set_language.js
@@ -2,7 +2,7 @@ const Logger = require("../../handlers/logger");
 const db      = require("quick.db");
 const data    = new db.table("suggestion_def")
 module.exports = {
-  name: "set_language", //the command name for the Slash Command
+  name: "language", //the command name for the Slash Command
   description: "Set the language that the bot uses to communitate with you. (Commands stay in English)", //the command description for Slash Command Overview
   cooldown: 5, //the cooldown for the command in seconds (max. 60) 
   memberpermissions: ["manage_server"], //Only allow members with specific Permissions to execute a Commmand [OPTIONAL]
